requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
Django==3.1.7
ibm-cos-sdk==2.10.0
ibm-cos-sdk-core==2.10.0
ibm-cos-sdk-s3transfer==2.10.0
idna==2.10
jmespath==0.10.0
python-dateutil==2.8.1
python-decouple==3.4
pytz==2021.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.4
